Use stdint.h in p00[1-3].c

+++ p001.c
@@ -1,9 +1,10 @@
+#include <stdint.h>
 #include <stdio.h>
 
 int main() {
-	int sum = 0;
-	for(int i = 1; i < 1000; i++)
-		if (i%3==0 || i%5==0)		
+	uint_fast32_t sum = 0;
+	for(uint_fast16_t i = 1; i < 1000; i++)
+		if (i%3==0 || i%5==0)
 			sum += i;
-	printf("%d\n", sum);
+	printf("%lu\n", sum);
 }
